+++ commands/listeners.go
@@ -2,8 +2,11 @@ package commands
 
 import (
 	"bonvoy/envoy"
-	"fmt"
+	"github.com/fatih/color"
+	"github.com/olekukonko/tablewriter"
+	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
+	"os"
 )
 
 type Listeners struct {
@@ -38,10 +41,24 @@ func (s *ListenersController) Run() error {
 	listeners, err := e.Listeners().Get()
 	if err != nil { return err }
 
-	fmt.Println("LISTENERS:")
-	fmt.Println("----------------------------------------------------------------------")
-	for _, listener := range listeners {
-		fmt.Printf("%s: %s\n", listener.Name, listener.TargetAddress)
+	color.Green("Listeners for "+s.ServiceName+" Envoy (PID "+cast.ToString(e.Pid)+")")
+	color.Green("----------------------------------------------------------------------")
+	table := tablewriter.NewWriter(os.Stdout)
+	table.SetHeader([]string{
+		"Name",
+		"Address",
+	})
+	table.SetBorder(false)
+	table.SetTablePadding("\t")
+	var d [][]string
+	for _, i := range listeners {
+		d = append(d, []string{
+			i.Name,
+			i.TargetAddress,
+		})
 	}
+	table.AppendBulk(d)
+	table.SetAlignment(tablewriter.ALIGN_LEFT)
+	table.Render()
 	return nil
 }